requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-genai==0.6.0
